Fixed a bug where enemies would always miss. Chnaged some display for combat log

+++ src/main/java/dungeon/model/characters/Ogre.java
@@ -45,8 +45,8 @@ public class Ogre extends Monster {
                 rs.getInt("MinDamage"),
                 rs.getInt("MaxDamage"),
                 rs.getInt("AttackSpeed"),
-                rs.getInt("HitChance"),
-                rs.getInt("HealChance"),
+                rs.getDouble("HitChance"),
+                rs.getDouble("HealChance"),
                 rs.getInt("MinHeal"),
                 rs.getInt("MaxHeal")
         );
