+++ src/main/java/dungeon/model/characters/Skeleton.java
@@ -43,8 +43,8 @@ public class Skeleton extends Monster {
                 rs.getInt("MinDamage"),
                 rs.getInt("MaxDamage"),
                 rs.getInt("AttackSpeed"),
-                rs.getInt("HitChance"),
-                rs.getInt("HealChance"),
+                rs.getDouble("HitChance"),
+                rs.getDouble("HealChance"),
                 rs.getInt("MinHeal"),
                 rs.getInt("MaxHeal")
         );
